task_2.3 fixes

+++ task_2.3.c
@@ -3,14 +3,14 @@
 #include <stdio.h>
 #include <time.h>
 
-int iter_fibonacci(int i) {
+unsigned long int iter_fibonacci(int i) {
     if (i == 0) {
         return 0;
     }
     if (i == 1) {
         return 1;
     }
-    int f1 = 0, f2 = 1, f_cur; 
+    unsigned long int f1 = 0, f2 = 1, f_cur; 
     for (int j = 2; j <= i; j++) {
         f_cur = f1 + f2;
         f1 = f2;
@@ -19,7 +19,7 @@ int iter_fibonacci(int i) {
     return f_cur;
 }
 
-int recurs_fibonacci(int i) {
+unsigned long int recurs_fibonacci(int i) {
     if (i == 0) {
         return 0;
     }
@@ -37,16 +37,16 @@ int main() {
     while (scanf("%d", &i) != EOF) {
         if (i >= 0) {
             clock_t start = clock();
-            int value = iter_fibonacci(i);
+            unsigned long int value = iter_fibonacci(i);
             clock_t end = clock();
             double duration = ((double) end - start) / CLOCKS_PER_SEC;
-            printf("fibonacci number (iteration function) = %d, took %f seconds to execute\n", value, duration);
+            printf("fibonacci number (iteration function) = %lu, took %f seconds to execute\n", value, duration);
             
             start = clock();
             value = recurs_fibonacci(i);
             end = clock();
             duration = ((double) end - start) / CLOCKS_PER_SEC;
-            printf("fibonacci number (recursive function) = %d, took %f seconds to execute\n", value, duration);
+            printf("fibonacci number (recursive function) = %lu, took %f seconds to execute\n", value, duration);
         } else {
             printf("incorrect number, try again\n");
         }
